# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mentorix Backend API
  version: 1.0.0
  description: >
    REST API for Mentorix: authentication, exercise catalog, training programs
    (draft/publish workflow), and admin operations.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://mentorix-backend.onrender.com
    description: Render dev
...
paths:
...
  /exercises:
    get:
      tags: [exercises]
      summary: List exercises (authenticated)
      operationId: listExercises
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SortBy"
        - $ref: "#/components/parameters/SortOrder"
        - name: q
          in: query
          schema:
            type: string
          description: Search in name and name_ru (ILIKE)
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/ExerciseType"
        - name: muscle_group
          in: query
          schema:
            $ref: "#/components/schemas/MuscleGroup"
        - name: difficulty
          in: query
          schema:
            $ref: "#/components/schemas/Difficulty"
        - name: equipment
          in: query
          schema:
            oneOf:
              - $ref: "#/components/schemas/Equipment"
              - type: string
                enum: [none]
          description: Filter by equipment; use `none` for exercises without equipment
      responses:
        "200":
          description: Paginated exercise list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [exercises]
      summary: Create exercise (admin only)
      description: >
        Shared catalog write. Any user with the `admin` role may create exercises.
        `created_by` records the acting admin; other admins may update or delete the entry later.
      operationId: createExercise
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseUpsert"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Admin role required

    delete:
      tags: [exercises]
      summary: Delete exercises by IDs (admin only)
      description: >
        Idempotent batch soft-delete. Sets `deleted_at` on existing exercises matching the given IDs.
        Any user with the `admin` role may delete exercises created by other admins.
        Soft-deleted exercises are hidden from list/get and cannot be added to programs.
        `deleted_count` may be less than the number of IDs if some were not found or already deleted.
        Duplicate IDs in the request are deduplicated before delete.
      operationId: deleteExercises
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseDeleteRequest"
      responses:
        "200":
          description: Exercises deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseDeleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Admin role required
  /exercises/{id}:
    parameters:
      - $ref: "#/components/parameters/ExerciseId"
    get:
      tags: [exercises]
      summary: Get exercise by ID (authenticated)
      operationId: getExercise
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Exercise
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "400":
          description: Invalid UUID
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Exercise not found

    put:
      tags: [exercises]
      summary: Update exercise (admin only)
      description: >
        Any user with the `admin` role may update any exercise in the shared catalog,
        including entries created by other admins.
      operationId: updateExercise
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseUpsert"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Admin role required
        "404":
          description: Exercise not found
...
  /programs/{id}/days/{day_id}/exercises:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/ProgramDayId"
    post:
      tags: [programs]
      summary: Add exercise to program day
      operationId: addProgramDayExercise
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramDayExerciseUpsert"
      responses:
        "201":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden (not owner and not admin)
        "404":
          description: Program or day not found
  /programs/{id}/days/{day_id}/exercises/{item_id}:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/ProgramDayId"
      - $ref: "#/components/parameters/ProgramDayExerciseId"
    put:
      tags: [programs]
      summary: Update exercise in program day
      operationId: updateProgramDayExercise
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramDayExerciseUpsert"
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden (not owner and not admin)
        "404":
          description: Program, day, or exercise item not found

    delete:
      tags: [programs]
      summary: Remove exercise from program day
      operationId: deleteProgramDayExercise
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden (not owner and not admin)
        "404":
          description: Program, day, or exercise item not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortBy:
      name: sort_by
      in: query
      schema:
        type: string
        enum:
          - name
          - name_ru
          - created_at
          - modified_at
          - difficulty
          - type
          - muscle_group
          - equipment
        default: name
    SortOrder:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    ExerciseId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProgramId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProgramDayId:
      name: day_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProgramDayExerciseId:
      name: item_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
    Unauthorized:
      description: Missing or invalid access token
...
  schemas:
...
    ExerciseListResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
        pagination:
          $ref: "#/components/schemas/Pagination"
    ExerciseDeleteRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    ExerciseDeleteResponse:
      type: object
      required: [deleted_count]
      properties:
        deleted_count:
          type: integer
          format: int64
          minimum: 0
    Pagination:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    Exercise:
      type: object
      required:
        - id
        - name
        - name_ru
        - created_by
        - created_by_name
        - modified_by
        - modified_at
        - created_at
        - type
        - muscle_group
        - description
        - description_ru
        - difficulty
        - video_url
        - preview_image_url
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        name_ru:
          type: string
        created_by:
          type: string
          format: uuid
        created_by_name:
          type: string
        modified_by:
          type: string
          format: uuid
        modified_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        equipment:
          $ref: "#/components/schemas/Equipment"
        type:
          $ref: "#/components/schemas/ExerciseType"
        muscle_group:
          $ref: "#/components/schemas/MuscleGroup"
        description:
          type: string
        description_ru:
          type: string
        difficulty:
          $ref: "#/components/schemas/Difficulty"
        video_url:
          type: string
        preview_image_url:
          type: string
    ExerciseUpsert:
      type: object
      required: [name, type, muscle_group, difficulty]
      properties:
        name:
          type: string
        name_ru:
          type: string
        equipment:
          $ref: "#/components/schemas/Equipment"
        type:
          $ref: "#/components/schemas/ExerciseType"
        muscle_group:
          $ref: "#/components/schemas/MuscleGroup"
        description:
          type: string
        description_ru:
          type: string
        difficulty:
          $ref: "#/components/schemas/Difficulty"
        video_url:
          type: string
        preview_image_url:
          type: string
    MuscleGroup:
      type: string
      enum:
        - compound
        - chest
        - back
        - legs
        - shoulders
        - arms
        - core
        - full_body
    ExerciseType:
      type: string
      enum:
        - strength
        - cardio
        - mixed_modal
        - intervals
        - stretching
        - metcon
        - skill_work
        - accessory
    Equipment:
      type: string
      enum:
        - barbell
        - dumbbells
        - kettlebell
        - pull_up_bar
        - squat_rack
        - rowing_machine
        - assault_bike
        - jump_rope
        - plyo_box
        - medicine_ball
        - wall_ball
        - resistance_bands
        - battle_ropes
        - gymnastic_rings
        - sandbag
        - sled
        - weight_plates
    Difficulty:
      type: string
      enum:
        - beginner
        - intermediate
        - advanced
        - expert
    ProgramStatus:
      type: string
      enum: [draft, published, archived]
    ProgramCategory:
      type: string
      enum:
        - weight_loss
        - muscle_gain
        - rehabilitation
        - endurance
        - functional
    Program:
      type: object
      required:
        - id
        - created_by
        - created_by_name
        - modified_by
        - status
        - name
        - name_ru
        - description
        - description_ru
        - preview_image_url
        - created_at
        - modified_at
      properties:
        id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        created_by_name:
          type: string
        modified_by:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ProgramStatus"
        name:
          type: string
        name_ru:
          type: string
        description:
          type: string
        description_ru:
          type: string
        category:
          $ref: "#/components/schemas/ProgramCategory"
          nullable: true
        difficulty:
          $ref: "#/components/schemas/Difficulty"
          nullable: true
        preview_image_url:
          type: string
        created_at:
          type: string
          format: date-time
        modified_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    ProgramDayExercise:
      type: object
      required:
        - id
        - exercise_id
        - exercise_name
        - exercise_name_ru
        - sort_order
        - instruction
        - created_at
      properties:
        id:
          type: string
          format: uuid
        exercise_id:
          type: string
          format: uuid
        exercise_name:
          type: string
        exercise_name_ru:
          type: string
        sort_order:
          type: integer
        sets:
          type: integer
          nullable: true
        reps:
          type: integer
          nullable: true
        weight_kg:
          type: number
          format: double
          nullable: true
        instruction:
          type: string
        created_at:
          type: string
          format: date-time
    ProgramDay:
      type: object
      required: [id, day_number, sort_order, exercises, created_at]
      properties:
        id:
          type: string
          format: uuid
        day_number:
          type: integer
        sort_order:
          type: integer
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/ProgramDayExercise"
        created_at:
          type: string
          format: date-time
    ProgramDetail:
      allOf:
        - $ref: "#/components/schemas/Program"
        - type: object
          required: [days]
          properties:
            days:
              type: array
              items:
                $ref: "#/components/schemas/ProgramDay"
...
    ProgramDayExerciseUpsert:
      type: object
      required: [exercise_id]
      description: >
        `sets` and `reps` are optional in draft; required with positive values on publish.
      properties:
        exercise_id:
          type: string
          format: uuid
        sets:
          type: integer
          minimum: 1
          nullable: true
        reps:
          type: integer
          minimum: 1
          nullable: true
        weight_kg:
          type: number
          format: double
          minimum: 0
        instruction:
          type: string